Create 07_Teams Power Users
Write a query to find the top 2 power users who sent the most messages on Microsoft Teams in August 2022. Display the IDs of these 2 users along with the total number of messages they sent. Output the results in descending count of the messages.

Use GROUP BY, EXTRACT clause for Year and Month and used the LIMIT operator to show the top 2 rows from the query