Use DISTINCT, CONCAT, SELF-JOIN
Assume you are given the following tables on Walmart transactions and products. Find the number of unique product combinations that are purchased in the same transaction.

For example, if there are 2 transactions where apples and bananas are bought, and another transaction where bananas and soy milk are bought, my output would be 2 to represent the 2 unique combinations.

Assumptions:
For each transaction, a maximum of 2 products is purchased.
You may or may not need to use the products table.